♻️ [REFACTOR] : js->jsx 파일 변경
