+++ src/pages/PostMessagePage/PostMessagePage.jsx
@@ -2,8 +2,8 @@ import React, { useState, useEffect } from 'react';
 import * as S from './PostMessagePage.style.js';
 import arrowDownIcon from '../../assets/icon/arrow_down.png';
 import arrowUpIcon from '../../assets/icon/arrow_top.png';
-import TextEditor from './components/TextEditor/TextEditor.js';
-import { COLORS } from '../../style/colorPalette';
+import TextEditor from './components/TextEditor/TextEditor.jsx';
+import { COLORS } from '../../style/colorPalette.js';
 import Header from '../../components/Common/Header/Header.jsx';
 import { useParams, useNavigate } from 'react-router-dom';
 import { PurpleButton } from '../../components/Common/PurpleButton/PurpleButton.jsx';
